requirements.txt
# Deployed dependencies
elasticsearch       >= 1.4.0
pyyaml              >= 3.10

# Development dependencies:
Nose == 1.1.2
Mock == 0.8.0
